Implemented appling rules logic to get and set command, fixed bug in lexeme parser with not truncated \r\n and not cleared vector of tokens, fixed bug with not passing commond to next handler. Added representations of get and set answers. BAsically, implemented basic behaviour for get and set commands.
Implemented appling rules logic to get and set command, fixed bug in lexeme parser with not truncated \r\n and not cleared vector of tokens, fixed bug with not passing commond to next handler. Added representations of get and set answers. BAsically, implemented basic behaviour for get and set commands.